Lista 1 Java

+++ java/Lista1/src/Exercicio3.java
@@ -1,21 +1,18 @@
 import java.util.Scanner;
-
 public class Exercicio3 {
-	public static void main(String[] args) {
+public static void main(String[] args) {
 		
 		Scanner read= new Scanner(System.in);
 		
-		double duracaoSeg;
+		int duracaoSeg;
 	
 		System.out.println("Tempo de duração do evento em segundos: ");
-		duracaoSeg=read.nextDouble();
+		duracaoSeg=read.nextInt();
 		
-		System.out.printf("%.2f horas, %.2f minutos, %.2f segundos ",(duracaoSeg/3600),((duracaoSeg%3600)/60),
+		System.out.printf("%d horas, %d minutos, %d segundos ",(duracaoSeg/3600),((duracaoSeg%3600)/60),
 		((duracaoSeg%3600)%60));
 		
 		read.close();
 	}
-	
-	
 
 }
